+++ CHANGELOG.md
@@ -14,6 +14,9 @@
 ### Changed
 - Pass the `height` prop to `VitessceGrid` so that the `WidthProvider` component can detect `height` changes.
 
+## Changed
+- Updated slider color for white slider with white theme.
+
 ## [0.1.4](https://www.npmjs.com/package/vitessce/v/0.1.4) - 2020-06-01
 
 ### Added